chore: clean up clippy warnings
Applied automatic clippy fixes and manual cleanup:
- Remove unused async from functions without await
- Fix empty line after doc comment
- Fix long literal lacking separators (0x12345678 -> 0x1234_5678)
- Fix doc list item overindentation

Reduced warnings from 280+ to 67. Remaining warnings are mostly:
- Casting warnings (need careful review for safety)
- Missing `# Errors` docs (pedantic/documentation)
- Too many function arguments (architectural refactoring needed)

+++ cosmic-connect-protocol/src/fs_utils.rs
@@ -50,7 +50,7 @@ pub async fn check_disk_space(path: impl AsRef<Path>, required_bytes: u64) -> Re
             Ok(stat) => {
                 // Available bytes = available blocks * fragment size
                 // fragment_size is the actual unit of allocation
-                let available_bytes = stat.blocks_available() as u64 * stat.fragment_size() as u64;
+                let available_bytes = stat.blocks_available() * stat.fragment_size();
 
                 debug!(
                     "Disk space check for {}: available={} bytes, required={} bytes",
